Moving the logic to an extension
